hardware_scripts/requirements.txt
# requirements.txt — Python dependencies for hardware_scripts/
#
# ── IMPORTANT: Install PyTorch with CUDA FIRST before running pip install ──
#
# RTX 3050 needs CUDA 12.1 build. Run this command first:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#
# Then install the rest:
#   pip install -r requirements.txt

# Serial communication with Arduino
pyserial==3.5

# HTTP requests to Convex
requests==2.32.3

# YOLO people counter (installs torch if not already installed — but without CUDA)
ultralytics==8.3.0

# OpenCV (used by YOLO and the detection window)
opencv-python==4.10.0.84

# Load environment variables from a .env file (optional, used in config.py)
python-dotenv==1.0.1
